articles: move ArticleForm to Views/, drop _form partial
Per the views/components rule (#4324): single-use components belong
in app/views/controllers/<controller>/. ArticleForm only renders for
the articles controller (new + edit), so it moves to:

  Components::ArticleForm
    -> Views::Controllers::Articles::Form
  app/components/article_form.rb
    -> app/views/controllers/articles/form.rb
  test/components/article_form_test.rb
    -> test/views/controllers/articles/form_test.rb

The `_form.html.erb` partial was a 1-line wrapper around the form
component; with the form now living in views/, the partial layer is
pure indirection. Deleted; new.html.erb and edit.html.erb render
the Phlex view directly.

Also tightens 4 test assertions that used the forbidden
`assert_html(html, "body", text: ...)` pattern to specific selectors
(`label[for='...']`, `.help-block`) — per the testing rule, every PR
that touches a test file is the right place to fix these.
