--- app/views/controllers/articles/_form.html.erb
@@ -1,1 +1,0 @@
-<%= render(Components::ArticleForm.new(@article)) %>
