+++ app/views/controllers/articles/edit.html.erb
@@ -3,4 +3,4 @@ add_edit_title(@article)
 add_context_nav(article_form_edit_tabs(article: @article))
 %>
 
-<%= render(partial: "articles/form") %>
+<%= render(Views::Controllers::Articles::Form.new(@article)) %>
