+++ app/views/controllers/articles/form.rb
@@ -1,0 +1,46 @@
+# frozen_string_literal: true
+
+module Views::Controllers::Articles
+  # Form for creating or editing articles. Rendered directly by the
+  # articles controller's `new.html.erb` and `edit.html.erb`.
+  # Articles support Textile markup for formatting.
+  class Form < ::Components::ApplicationForm
+    def view_template
+      super do
+        render_title_field
+        render_body_field
+        submit(:SUBMIT.t, center: true)
+      end
+    end
+
+    private
+
+    def render_title_field
+      text_field(:title, label: "#{:article_title.t}:",
+                         data: { autofocus: true }) do |field_component|
+        field_component.with_append do
+          help_block do
+            "#{:form_article_title_help.t} #{:field_textile_link.t}"
+          end
+        end
+      end
+    end
+
+    def render_body_field
+      textarea_field(:body, label: "#{:article_body.t}:",
+                            rows: 10) do |field_component|
+        field_component.with_append do
+          help_block do
+            :field_textile_link.t
+          end
+        end
+      end
+    end
+
+    def help_block
+      # rubocop:disable Rails/OutputSafety
+      div(class: "help-block") { raw(yield.html_safe) }
+      # rubocop:enable Rails/OutputSafety
+    end
+  end
+end
